first matlab files
this is the first draft of the code for the analysis. it creates a lot of the plots we looked at for the TRB paper. There still is more to be added to fill out the analysis for a single driver. And then we need to do aggregate analysis to see trends across sets of drivers.